Attempt to exclude binary cookies from backup
